requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python-headless
numpy
torch
ultralytics
paddlepaddle
paddleocr
pyclipper
shapely
scikit-image
imgaug
google-generativeai
Pillow
python-dotenv
pydantic
pydantic-settings
python-logging-loki
requests
openai